requirements.txt
Flask==2.2.2
newspaper3k==0.2.8
torch==2.4.1
transformers==4.45.2